refactor persistence API
