+++ .vscode/settings.json
@@ -6,8 +6,11 @@
     "imagesdir",
     "mariotoffia",
     "nocng",
+    "Nygatan",
     "omitempty",
+    "Östra",
     "Persistable",
+    "persistencemodel",
     "processmodel",
     "stretchr",
     "strlogger",
